requirements.txt
asgiref==3.4.1
black==21.10b0
click==8.0.3
Django==3.2.9
django-filter==21.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
Markdown==3.3.4
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumbers==8.12.36
platformdirs==2.4.0
pytz==2021.3
regex==2021.11.2
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
